Ana/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
argon2-cffi
pydantic
